Add integration test 1, positive (#64)
* Add script generate LCM

* Add integration test 1

+++ src/main/java/com/group23/decide/Point.java
@@ -7,13 +7,13 @@ import java.lang.Math;
  */
 public class Point {
 
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     /**
      * The constructor for a Point takes two arguments, x and y and returns an Object of type Point.
      */
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
@@ -21,28 +21,28 @@ public class Point {
     /**
      * Getter for x
      */
-    public int getX() {
+    public double getX() {
         return x;
     }
 
     /**
      * Getter for y
      */
-    public int getY() {
+    public double getY() {
         return y;
     }
 
     /**
      * Setter for x
      */
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
     /**
      * Setter for y
      */
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
